AWS S3 Signature V4 - Fix signature calculation for similar headers
Boto calculated a incorrect signature for requests containing similar
headers as it sorted the headers not by the header name but by the
complete header.

An example is the multipart upload part copy with range api. Boto
calculates its signature using this order because - sorts before :
    x-amz-copy-source-range:bytes=5-15
    x-amz-copy-source:multipart-1426209034/source

The order is incorrect and should be the other way around. The new
algorithm uses only the header name to sort the headers.

Two tests are added which fail with the old algorithm.